refactor(WelcomeTemplate): make WelcomeTemplate file entrypoint for imports and handle platform specific logic internally
