+++ app/components/platformSpecific/WelcomeTemplate.electron.tsx
@@ -1,0 +1,3 @@
+import { remote } from 'electron'
+
+export const onExit = () => remote.app.quit()
