+++ app/components/platformSpecific/WelcomeTemplate.web.tsx
@@ -1,0 +1,1 @@
+export const onExit = () => true
